% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_GTAPauto.R
\name{.apply_filters}
\alias{.apply_filters}
\title{Apply Filters to GTAP Data (Internal)}
\usage{
.apply_filters(
  data,
  region_select = NULL,
  experiment_select = NULL,
  sector_select = NULL
)
}
\arguments{
\item{data}{List of data frames or single data frame to filter.}

\item{region_select}{Character vector of regions to include.}

\item{experiment_select}{Character vector of experiments to include.}

\item{sector_select}{Character vector of sectors to include.}
}
\value{
Filtered data in the same structure as input.
}
\description{
Applies region, experiment, and sector filters to GTAP data structures.
}
\seealso{
\code{\link{process_gtap_data}}
}
\author{
Pattawee Puangchit
}
\keyword{internal}
