% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTAP_Dynamic.R
\name{.auto_GTAPRd_info}
\alias{.auto_GTAPRd_info}
\title{Add Mapping Information to GTAP Data}
\usage{
.auto_GTAPRd_info(data_list, external_map = NULL)
}
\arguments{
\item{data_list}{A data frame or list containing GTAP data structures.}

\item{external_map}{Data frame with mapping information. Must contain columns:
"Input", "Case", and "Period".}
}
\value{
The input data structure with added Case and Period columns.
}
\description{
Adds Case and Period information to GTAP data based on a provided mapping.
}
\author{
Pattawee Puangchit
}
\keyword{internal}
