% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_plot.R
\name{.calculate_plot_dimensions}
\alias{.calculate_plot_dimensions}
\title{Calculate Plot Dimensions}
\usage{
.calculate_plot_dimensions(data, panel_layout)
}
\arguments{
\item{data}{A data frame containing the relevant plotting data.}

\item{panel_layout}{A list containing 'rows' and 'cols' specifying the panel layout.}
}
\value{
A list with 'width' and 'height' specifying the calculated plot dimensions in inches.
}
\description{
Determines appropriate width and height for a plot based on the panel layout.
}
\seealso{
\code{\link{comparison_plot}}, \code{\link{detail_plot}}, \code{\link{stack_plot}}
}
\author{
Pattawee Puangchit
}
\keyword{internal}
