% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{.calculate_stack_totals}
\alias{.calculate_stack_totals}
\title{Calculate Stack Totals}
\usage{
.calculate_stack_totals(data, x_axis_from, panel_var)
}
\arguments{
\item{data}{Data frame containing stack components.}

\item{x_axis_from}{Column for x-axis categories.}

\item{panel_var}{Column for panel facets.}
}
\value{
Data frame with calculated totals.
}
\description{
Internal function to calculate totals for stack components.
}
\keyword{internal}
