% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.coalesce}
\alias{.coalesce}
\title{Coalesce Two Values}
\usage{
.coalesce(x, y)
}
\arguments{
\item{x}{First value to check.}

\item{y}{Fallback value if `x` is NULL.}
}
\value{
The first non-NULL value. If `x` is NULL, returns `y`.
}
\description{
Returns the first non-NULL value from the given inputs.
}
\keyword{internal}
