% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTAP_Dynamic.R
\name{.create_agg_mapping}
\alias{.create_agg_mapping}
\title{Create Aggregation Mapping Structure}
\usage{
.create_agg_mapping(
  mapping_file,
  dimension_col = "Dimension",
  item_col = "GTAPName",
  group_col = "Aggregate"
)
}
\arguments{
\item{mapping_file}{A data frame or path to a file with mapping information.}

\item{dimension_col}{Character. Name of column with dimension information.}

\item{item_col}{Character. Name of column with original GTAP names.}

\item{group_col}{Character. Name of column with aggregation group names.}
}
\value{
A nested list with aggregation mappings.
}
\description{
Creates a hierarchical mapping structure for region/sector aggregation.
}
\author{
Pattawee Puangchit
}
\keyword{internal}
