% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{.filter_top_impact_values_detail}
\alias{.filter_top_impact_values_detail}
\title{Filter Top Impact Values (Internal)}
\usage{
.filter_top_impact_values_detail(
  data,
  top_impact,
  group_col,
  panel_var,
  x_axis_from,
  variable_col,
  unit_col
)
}
\arguments{
\item{data}{A data frame or list to filter.}

\item{top_impact}{Numeric. Number of top impacts to display.}

\item{group_col}{Character. Column for grouping data.}

\item{panel_var}{Character. Column for panel facets.}

\item{x_axis_from}{Character. Column for x-axis categories.}

\item{variable_col}{Character. Column for variable identification.}

\item{unit_col}{Character. Column for unit information.}
}
\value{
A filtered data frame or list.
}
\description{
Filter data to show only top impactful values from GTAP data for internal use.
}
\seealso{
\code{\link{detail_plot}}
}
\author{
Pattawee Puangchit
}
\keyword{internal}
