% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{.filter_top_impact_values_stack}
\alias{.filter_top_impact_values_stack}
\title{Filter Top Impact Values for Stack Plots (Internal)}
\usage{
.filter_top_impact_values_stack(
  data,
  total_data,
  top_impact,
  group_col,
  panel_var,
  x_axis_from,
  variable_col,
  unit_col,
  stack_value_from
)
}
\arguments{
\item{data}{A data frame to filter.}

\item{total_data}{A data frame with total values.}

\item{top_impact}{Numeric. Number of top impacts to display.}

\item{group_col}{Character. Column for grouping data.}

\item{panel_var}{Character. Column for panel facets.}

\item{x_axis_from}{Character. Column for x-axis categories.}

\item{variable_col}{Character. Column for variable identification.}

\item{unit_col}{Character. Column for unit information.}

\item{stack_value_from}{Character. Column for stack components.}
}
\value{
A filtered data frame.
}
\description{
Filter data for stack plots to show only top impactful values from GTAP data for internal use.
}
\seealso{
\code{\link{stack_plot}}
}
\author{
Pattawee Puangchit
}
\keyword{internal}
