% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_plot.R
\name{.get_title_mapping}
\alias{.get_title_mapping}
\title{Generate Title Mapping for Plot Labels (Internal)}
\usage{
.get_title_mapping(data, description_as_title = TRUE)
}
\arguments{
\item{data}{A data frame or a list of data frames containing the dataset.}

\item{description_as_title}{Logical. If \code{TRUE}, prioritizes the \code{Description} column for titles.
If \code{FALSE}, using \code{Variable} as title name.}
}
\value{
A named vector where variable names map to their corresponding titles.
}
\description{
Determines appropriate title names for plotting based on available columns in the dataset.
If a \code{Description} column exists and \code{description_as_title} is \code{TRUE}, it uses \code{Description} as titles.
Otherwise, it checks for \code{PlotTitle} or falls back to using \code{Variable} names.
}
\author{
Pattawee Puangchit
}
\keyword{internal}
