% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTAP_Dynamic.R
\name{.gtap_rd_agg}
\alias{.gtap_rd_agg}
\title{Aggregate Regions and Sectors in GTAP Data}
\usage{
.gtap_rd_agg(data_list, agg_mapping, calculation = "+", add_world = TRUE)
}
\arguments{
\item{data_list}{A data frame or list containing GTAP data to be aggregated.}

\item{agg_mapping}{A list containing aggregation mappings.}

\item{calculation}{Character. Operation for aggregation.}

\item{add_world}{Logical. If TRUE, adds a "World" aggregate for regions.}
}
\value{
The input data structure with aggregated values added.
}
\description{
Aggregates regions, sectors, or other dimensions in GTAP data.
}
\author{
Pattawee Puangchit
}
\keyword{internal}
