% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTAP_Dynamic.R
\name{.gtap_rd_dev}
\alias{.gtap_rd_dev}
\title{Calculate Deviations Between Policy and Base Scenarios}
\usage{
.gtap_rd_dev(data_list, base = "base", policy = "pol", calculation = "-")
}
\arguments{
\item{data_list}{A data frame or list containing GTAP data with Case column.}

\item{base}{Character. The identifier for baseline scenario in the Case column.}

\item{policy}{Character. The identifier for policy scenario in the Case column.}

\item{calculation}{Character. Operation for calculating deviations.}
}
\value{
The input data structure with deviation values added as a new case.
}
\description{
Computes deviations between policy and baseline scenarios in GTAP data.
}
\author{
Pattawee Puangchit
}
\keyword{internal}
