% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_plot.R
\name{.prepare_data_source}
\alias{.prepare_data_source}
\title{Prepare Data Source from List or Data Frame}
\usage{
.prepare_data_source(data, required_columns = NULL)
}
\arguments{
\item{data}{A data frame or list of data frames.}

\item{required_columns}{Optional vector of column names that must be present.}
}
\value{
A data frame that contains the required columns.
}
\description{
Finds a suitable data frame from a list or returns the input data frame.
}
\seealso{
\code{\link{comparison_plot}}, \code{\link{detail_plot}}, \code{\link{stack_plot}}
}
\author{
Pattawee Puangchit
}
\keyword{internal}
