% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.select_top_impact}
\alias{.select_top_impact}
\title{Select Top Impact Items}
\usage{
.select_top_impact(data, top_n, group_by_col)
}
\arguments{
\item{data}{Data frame containing the data to filter.}

\item{top_n}{Number of top items to select.}

\item{group_by_col}{Column name to group by when selecting top items.}
}
\value{
Filtered data frame with only top impact items.
}
\description{
Selects the top N items based on absolute value of the impact.
}
\keyword{internal}
