% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTAP_Dynamic.R
\name{dynamic_input_name}
\alias{dynamic_input_name}
\title{Generate Dynamic Input Mapping for GTAP Simulations}
\usage{
dynamic_input_name(
  type = "prefix",
  base = NULL,
  pol = NULL,
  base_rerun = NULL,
  other = NULL,
  pattern = NULL,
  increment = 1,
  separator = "-",
  period_pattern = FALSE,
  period_prefix = "yr_",
  output = "Input_map"
)
}
\arguments{
\item{type}{Naming pattern}

\item{base}{Identifiers for baseline cases}

\item{pol}{Identifiers for policy cases}

\item{base_rerun}{Identifiers for baseline rerun cases}

\item{other}{Identifiers for other case types}

\item{pattern}{Pattern to be combined with case identifiers}

\item{increment}{Step size for numeric sequences}

\item{separator}{Separator between pattern and case identifier}

\item{period_pattern}{Creates period ranges between patterns}

\item{period_prefix}{Prefix for period names}

\item{output}{Name for result variable}
}
\value{
A data frame mapping experiments to cases and periods

A data frame mapping experiments to cases and periods.
}
\description{
Creates a mapping data frame of simulation experiments to cases and periods
for dynamic GE model analysis.
}
\author{
Pattawee Puangchit
}
