% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{rename_GTAP_bilateral}
\alias{rename_GTAP_bilateral}
\title{Rename GTAP Bilateral Trade Columns}
\usage{
rename_GTAP_bilateral(data)
}
\arguments{
\item{data}{Data structure containing GTAP bilateral trade data.}
}
\value{
The same data structure with renamed bilateral trade columns.
}
\description{
Renames bilateral trade columns in GTAP data to standardized names,
ensuring consistency in regional trade flows.
}
\examples{
# Load Sample Data:
sl4_data1 <- HARplus::load_sl4x(system.file("extdata/in", "EXP1.sl4", 
                                package = "GTAPViz"))
# Get Data by Variable Name
sl4_data1 <- HARplus::get_data_by_var("qxs",sl4_data1)

# Rename bilateral trade columns in a GTAP dataset
gtap_data <- rename_GTAP_bilateral(sl4_data1)

}
\seealso{
\code{\link{add_mapping_info}}, \code{\link{convert_units}}, \code{\link{sort_plot_data}}
}
\author{
Pattawee Puangchit
}
