% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{scalar_table}
\alias{scalar_table}
\title{Create Scalar Result Table}
\usage{
scalar_table(
  data,
  vars = NULL,
  output_dir = NULL,
  workbook_name = "scalar_results",
  sheet_name = "Results"
)
}
\arguments{
\item{data}{A data frame or list containing scalar data, where each observation consists of
an experiment name, variable name, and corresponding value.}

\item{vars}{Optional character vector of variable names to include in the table.
If \code{NULL}, all available variables in the dataset will be used.}

\item{output_dir}{Optional character. Directory where the generated table will be saved
as an Excel file. If \code{NULL}, no file will be created.}

\item{workbook_name}{Optional character. Name of the output Excel workbook. Default is \code{"scalar_results"}.}

\item{sheet_name}{Optional character. Name of the worksheet in the Excel file. Default is \code{"Results"}.}
}
\value{
A formatted data frame in wide format with variables as rows and experiments as columns.
If \code{output_dir} is provided, the function also saves the output to an Excel file.
}
\description{
Generates a formatted table summarizing scalar results from a given dataset.
The function transforms long-format data into a wide format with scenarios as columns,
making comparisons across experiments more accessible.
}
\details{
\itemize{
\item If \code{data} is a list and not a data frame, the function attempts to combine it into a data frame.
\item The table is formatted to ensure numerical values are rounded to two decimal places.
\item If \code{output_dir} is specified, the table is exported as an Excel file with proper formatting.
\item Required columns in \code{data} are \code{"Experiment"}, \code{"Variable"}, and \code{"Value"}.
}
}
\examples{
\dontrun{
sl4_data1 <- HARplus::get_data_by_var(c("globalcgds", "vgdpwld", "WEV"), sl4_data)
scalar_table <- scalar_table(
  data = sl4_data1,
  vars = "vgdpwld",
  output_dir = temp_dir,
  workbook_name = "global_indicators",
  sheet_name = "Results"
)
}
}
\seealso{
\code{\link{report_table}}, \code{\link{decomp_table}}
}
\author{
Pattawee Puangchit
}
