% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{convert_units}
\alias{convert_units}
\title{Convert Multiple Units in Nested Data Structures}
\usage{
convert_units(
  data,
  change_unit_from,
  change_unit_to,
  adjustment,
  value_col = "Value",
  unit_col = "Unit"
)
}
\arguments{
\item{data}{A data structure (list, data.frame, or nested combination)}

\item{change_unit_from}{Character vector of units to change (case-insensitive)}

\item{change_unit_to}{Character vector of new units (same length as change_unit_from)}

\item{adjustment}{Character vector of operations or numeric vector (same length as change_unit_from)}

\item{value_col}{Column name containing values to adjust (default: "Value")}

\item{unit_col}{Column name containing unit information (default: "Unit")}
}
\value{
Data structure with same format as input but with adjusted values and units
}
\description{
Convert Multiple Units in Nested Data Structures
}
\examples{
\dontrun{
# Load example data
har_data <- HARplus::load_harx(system.file("extdata/in", "EXP1.sl4",
                                          package = "GTAPViz"))

har_data <- convert_units(har_data,
change_unit_from = c("million USD"),
change_unit_to = c("billion USD"),
adjustment = c("/1000"))
}

}
\author{
Pattawee Puangchit
}
