% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{detail_plot}
\alias{detail_plot}
\title{Generate Detailed Plot}
\usage{
detail_plot(
  data,
  plot_var = NULL,
  y_axis_from,
  figure_separate_by,
  title_prefix = "",
  title_suffix = "",
  panel_rows = NULL,
  panel_cols = NULL,
  positive_color = "#2E8B57",
  negative_color = "#CD5C5C",
  output_dir = NULL,
  top_impact = NULL,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{data}{A data frame containing the full dataset.}

\item{plot_var}{A data frame with "Variable" and "PlotTitle" columns specifying which variables to plot and their titles.
If NULL, all variables in the data will be plotted with default titles.}

\item{y_axis_from}{Character vector. Column names to use for the y-axis.}

\item{figure_separate_by}{Character. Column used to separate different plots.}

\item{title_prefix}{Optional character string to prepend to the title.}

\item{title_suffix}{Optional character string to append to the title.}

\item{panel_rows}{Optional numeric. Number of panel rows. If left blank, it will be automatically adjusted.}

\item{panel_cols}{Optional numeric. Number of panel columns. If left blank, it will be automatically adjusted.}

\item{positive_color}{Character. Color for positive values.}

\item{negative_color}{Character. Color for negative values.}

\item{output_dir}{Optional character. Directory to save the output plot.}

\item{top_impact}{Optional numeric. Number of top impact factors to include. If left blank, all factors will be included.}

\item{width}{Optional numeric. Width of the output plot. If left blank, it will be automatically adjusted.}

\item{height}{Optional numeric. Height of the output plot. If left blank, it will be automatically adjusted.}
}
\value{
A \code{ggplot} object or NULL if no plots could be generated.
}
\description{
Creates a detailed plot reporting all regions or sectors depending on the variable selection.
Uses a separate dataframe to specify variables to plot and their titles.
}
