% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTAPauto.R
\name{.ask_confirmation}
\alias{.ask_confirmation}
\title{Ask for User Confirmation}
\usage{
.ask_confirmation(prompt)
}
\arguments{
\item{prompt}{A character string specifying the message to display to the user.}
}
\value{
A logical value: TRUE if the user types "y" (case-insensitive), FALSE if "n".
}
\description{
Prompts the user for confirmation by displaying a message and reading input from the console. Returns TRUE if the user confirms with 'y', otherwise FALSE.
}
\author{
Pattawee Puangchit
}
\keyword{internal}
