% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_plot.R
\name{.calculate_panel_layout}
\alias{.calculate_panel_layout}
\title{Calculate Optimal Panel Layout}
\usage{
.calculate_panel_layout(
  data,
  panel_rows = NULL,
  panel_cols = NULL,
  panel_var = "Experiment"
)
}
\arguments{
\item{data}{A data frame containing the relevant plotting data.}

\item{panel_rows}{Optional. Number of rows for the panel layout.}

\item{panel_cols}{Optional. Number of columns for the panel layout.}

\item{panel_var}{Column name to use for determining unique panels.}
}
\value{
A list with 'rows' and 'cols' specifying the calculated panel layout.
}
\description{
Determines the optimal panel layout (rows and columns) for plotting based on the number of panels needed.
}
\seealso{
\code{\link{comparison_plot}}, \code{\link{detail_plot}}, \code{\link{stack_plot}}
}
\author{
Pattawee Puangchit
}
\keyword{internal}
