% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_plot.R
\name{.generate_color_palette}
\alias{.generate_color_palette}
\title{Generate Color Palette for Positive/Negative Comparison (Internal)}
\usage{
.generate_color_palette(positive_color = "#2E8B57", negative_color = "#CD5C5C")
}
\arguments{
\item{positive_color}{Character. Hex code for the positive color.}

\item{negative_color}{Character. Hex code for the negative color.}
}
\value{
A named vector containing hex codes for different shades.
}
\description{
Generates a color palette with shades for positive, negative, and neutral values.
}
\author{
Pattawee Puangchit
}
\keyword{internal}
