% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.generate_comparison_colors}
\alias{.generate_comparison_colors}
\title{Generate Comparison Colors (Internal)}
\usage{
.generate_comparison_colors(
  data,
  color_tone = NULL,
  compare_by_x_axis = FALSE,
  x_axis_from = NULL
)
}
\arguments{
\item{data}{A data frame containing the relevant plotting data.}

\item{color_tone}{Character. Base color for generating shades.}

\item{compare_by_x_axis}{Logical. If \code{TRUE}, colors are generated based on \code{Experiment} values.}

\item{x_axis_from}{Character. Column name used when \code{compare_by_x_axis = TRUE}.}
}
\value{
A vector of colors in hexadecimal format.
}
\description{
Generates a color palette for comparison plots based on a base color tone.
}
\author{
Pattawee Puangchit
}
\keyword{internal}
