% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_plot.R
\name{.generate_stack_colors}
\alias{.generate_stack_colors}
\title{Generate Colors for Stacked Bar Components}
\usage{
.generate_stack_colors(
  data,
  stack_value_from,
  color_tone = NULL,
  palette_type = "qualitative"
)
}
\arguments{
\item{data}{A data frame containing the stack value column.}

\item{stack_value_from}{Column name containing the stack categories.}

\item{color_tone}{Optional base color to influence the palette.}

\item{palette_type}{Character. Optional color palette by default is qualitative.}
}
\value{
A named vector of colors for each stack component.
}
\description{
Generates a color palette for stacked bar components, ensuring good contrast between items.
}
\seealso{
\code{\link{stack_plot}}
}
\author{
Pattawee Puangchit
}
\keyword{internal}
