% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTAPauto.R
\name{.output_format}
\alias{.output_format}
\title{Determine Output Formats}
\usage{
.output_format()
}
\value{
A character vector of output format names.
}
\description{
Returns a character vector of output format names for which the corresponding global variable is set to "yes".
}
\seealso{
\code{\link{process_gtap_data}}
}
\author{
Pattawee Puangchit
}
\keyword{internal}
