% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{.process_detail_data}
\alias{.process_detail_data}
\title{Prepare Detail Data for Wide-Format Report (Internal)}
\usage{
.process_detail_data(
  df,
  wide_col,
  group_cols,
  rename_mapping,
  total_column,
  decimal
)
}
\arguments{
\item{df}{A data frame to be transformed.}

\item{wide_col}{Character. The column to pivot into wide format.}

\item{group_cols}{Character vector. The grouping columns to retain.}

\item{rename_mapping}{Named list for renaming grouping columns.}

\item{total_column}{Logical. If \code{TRUE}, adds a "Total" column. Default is \code{FALSE}.}

\item{decimal}{Numeric. Number of decimal places to round numeric columns. Default is \code{2}.}

\item{col_name_by_description}{Logical. If \code{TRUE}, uses the "Description" column for naming
pivoted columns. Also, if \code{"Variable"} stays in rows, we replace its values with description.}

\item{sub_column_name}{Logical. If \code{TRUE}, appends original variable names in parentheses. Default is \code{FALSE}.}
}
\value{
A data frame in wide format, with optionally added totals, renamed columns,
sorted grouping columns, and numeric columns rounded to the specified decimal places.
}
\description{
Converts a long-format data frame into a wide-format structure, optionally calculating
totals, renaming columns, handling descriptions, and sorting the final table.
\strong{Also supports the special case} where \code{Variable} is \emph{not} pivoted and remains in the row
dimension (e.g. \code{group_by = "Variable"}). In that scenario, if \code{col_name_by_description = TRUE},
it replaces the actual values of \code{"Variable"} with a more descriptive label (and optionally
adds the original variable name in parentheses if \code{sub_column_name = TRUE}).
}
\author{
Pattawee Puangchit
}
\keyword{internal}
