% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTAPauto.R
\name{gtap_macros_data}
\alias{gtap_macros_data}
\title{Extract and Aggregate Scalar Macroeconomic Variables}
\usage{
gtap_macros_data(
  input_dir = NULL,
  experiment_names = NULL,
  output_dir = NULL,
  output_formats = NULL,
  subtotal_level = FALSE,
  select_var = NULL
)
}
\arguments{
\item{input_dir}{Character. Directory containing SL4 files.}

\item{experiment_names}{Character vector. List of experiment names corresponding to SL4 files.}

\item{output_dir}{Character (optional). Directory to save exported data.}

\item{output_formats}{Character vector (optional). List of output formats (e.g., "csv", "xlsx").}

\item{subtotal_level}{Logical. Whether to include subtotal levels in the processed data.}

\item{select_var}{Character vector (optional). List of specific variable names to filter from the final result. If NULL, all variables are returned.}
}
\value{
A sorted dataframe containing processed GTAP macro data.
}
\description{
Extracts scalar macroeconomic variables from multiple SL4 datasets and aggregates them into a structured data frame.
}
\examples{
\dontrun{
# Method 1: Extract all variables
input_dir <- paste0(input.folder)
experiment <- c("EXP1", "EXP2") # File name (.sl4)
Macros <- gtap_macros_data(input_dir, experiment_names = experiment,
subtotal_level = FALSE)

# Method 2: Filter specific variables
Macros <- gtap_macros_data(input_dir, experiment_names = experiment,
                          select_var = c("qgdp", "pop", "gdpexp"))
}

}
\seealso{
\code{\link{add_mapping_info}}, \code{\link{auto_gtap_data}}
}
\author{
Pattawee Puangchit
}
