% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTAPauto.R
\name{gtap_macros_data}
\alias{gtap_macros_data}
\title{Extract and Aggregate Scalar Macroeconomic Variables}
\usage{
gtap_macros_data(
  select_var = NULL,
  experiment = NULL,
  input_path = NULL,
  output_path = NULL,
  output_formats = NULL,
  subtotal_level = FALSE
)
}
\arguments{
\item{select_var}{Character vector (optional). List of specific variable names to filter from the final result. If NULL, all variables are returned.}

\item{experiment}{Character vector. List of experiment names corresponding to SL4 files.}

\item{input_path}{Character. Path to the directory containing SL4 files.}

\item{output_path}{Character (optional). Directory to save exported data.}

\item{output_formats}{Character vector (optional). List of output formats (e.g., "csv", "xlsx").}

\item{subtotal_level}{Logical. Whether to include subtotal levels in the processed data.}
}
\value{
A sorted data frame containing processed GTAP macro data.
}
\description{
Extracts scalar macroeconomic variables from multiple SL4 datasets and aggregates them into a structured data frame.
}
\examples{

# Input Path:
input_path <- system.file("extdata/in", package = "GTAPViz")

# GTAP Macro Variables from 2 .sl4 Files named (EXP1, EXP2)
# Note: No need to add .sl4 to the experiment name
gtap_macro <- gtap_macros_data(NULL, experiment = c("EXP1", "EXP2"),
                               input_path = input_path, subtotal_level = FALSE)

}
\seealso{
\code{\link{add_mapping_info}}, \code{\link{auto_gtap_data}}
}
\author{
Pattawee Puangchit
}
