% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{pivot_table_with_filter}
\alias{pivot_table_with_filter}
\title{Export Data as an Excel Pivot Table}
\usage{
pivot_table_with_filter(
  data,
  filter = NULL,
  rows = NULL,
  cols = NULL,
  data_fields = "Value",
  raw_sheet_name = "RawData",
  pivot_sheet_name = "PivotTable",
  dims = "A5",
  export = TRUE,
  output_path = getwd(),
  workbook_name = "GTAP_PivotTable.xlsx"
)
}
\arguments{
\item{data}{Data frame. The dataset to be exported.}

\item{filter}{Character vector (optional). Columns to be used as filter fields in the pivot table.}

\item{rows}{Character vector (optional). Columns to be used as row fields in the pivot table.}

\item{cols}{Character vector (optional). Columns to be used as column fields in the pivot table.}

\item{data_fields}{Character. The data field(s) to be summarized in the pivot table (default: `"Value"`).}

\item{raw_sheet_name}{Character. Name of the sheet containing raw data (default: `"RawData"`).}

\item{pivot_sheet_name}{Character. Name of the sheet containing the pivot table (default: `"PivotTable"`).}

\item{dims}{Character. Cell reference where the pivot table starts (default: `"A3"`).}

\item{export}{Logical. Whether to save the Excel file (default: `TRUE`).}

\item{output_path}{Character. Directory where the file should be saved (default: current working directory).}

\item{workbook_name}{Character. Name of the output Excel file (default: `"GTAP_PivotTable.xlsx"`).}
}
\value{
An excel workbook object containing both raw data and the pivot table.
}
\description{
Exports a dataset to an Excel file with both raw data and a generated pivot table.
}
\details{
This function creates an Excel workbook with:
- A raw data sheet (`raw_sheet_name`) containing the provided dataset.
- A pivot table sheet (`pivot_sheet_name`) generated based on specified row, column, and data fields.

If `export = TRUE`, the function saves the workbook to the specified `output_path`.
}
\examples{
\donttest{
# Input Path
input_path <- system.file("extdata/in", package = "GTAPViz")

# Prepares data from SL4 files for EXP1 and EXP2
auto_gtap_data(
  experiment = c("EXP1", "EXP2"),
  input_path = input_path,
  subtotal_level = FALSE,
  process_sl4_vars = c("qo", "qgdp"),
  process_har_vars = FALSE,
  mapping_info = "GTAPv7",
  plot_data = TRUE
)

# Generate comparison table
pivot_table_with_filter(
  data = sl4.plot.data[["2D"]][["Sector"]],
  filter = c("Variable", "Unit"),
  rows = c("Region", "Sector"),
  cols = c("Experiment"),
  data_fields = "Value",
  raw_sheet_name = "Raw_Data",
  pivot_sheet_name = "Sector_Pivot",
  export = TRUE,
  output_path = tempdir(),
  workbook_name = "Sectoral_Impact_Analysis.xlsx"
)
}

}
\author{
Pattawee Puangchit
}
