% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{rename_GTAP_bilateral}
\alias{rename_GTAP_bilateral}
\title{Rename GTAP Bilateral Trade Columns}
\usage{
rename_GTAP_bilateral(data)
}
\arguments{
\item{data}{Data structure containing GTAP bilateral trade data}
}
\value{
Data structure with renamed bilateral trade columns
}
\description{
Renames bilateral trade columns in GTAP data output
}
\examples{
\dontrun{
# Load Sample Data
sl4_data <- HARplus::load_sl4x(system.file("extdata/in", "EXP1-WEL.sl4",
                                          package = "GTAPViz"))
sl4_data <- rename_GTAP_bilateral(df)
}

}
\author{
Pattawee Puangchit
}
