% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{rename_GTAP_bilateral}
\alias{rename_GTAP_bilateral}
\title{Rename GTAP Bilateral Trade Columns}
\usage{
rename_GTAP_bilateral(data)
}
\arguments{
\item{data}{List or data frame. Output from get_data_by_var, get_data_by_dims, or group_data_by_dims.}
}
\value{
Data with renamed columns, maintaining the original structure type.
}
\description{
Renames bilateral trade columns in GTAP data output, specifically targeting REG columns
to Source and Destination in data frames with COMM\emph{REG}REG dimension structure.
}
\examples{
\dontrun{
# Load Sample Data
sl4_data <- HARplus::load_sl4x(system.file("extdata", "TAR10.sl4", package = "HARplus"))

# For direct data frame output
df <- HARplus::get_data_by_dims("COMM*REG*REG", sl4_data)
df_renamed <- rename_GTAP_bilateral(df)
}

}
\author{
Your Name
}
