% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{rename_value}
\alias{rename_value}
\title{Rename Values in a Column}
\usage{
rename_value(data, column_name = NULL, mapping.file)
}
\arguments{
\item{data}{Data structure (data frame, list, or nested combination).}

\item{column_name}{Character. Column to modify. If `NULL`, the function extracts it from `mapping.file`.}

\item{mapping.file}{Data frame with `"OldName"` and `"NewName"` columns for renaming.}
}
\value{
The same data structure with specified values replaced.
}
\description{
Replaces specific values in a column based on a provided mapping file.
Supports renaming across nested data structures and preserves factor levels.
}
\examples{
# Load Sample Data:
sl4_data1 <- HARplus::load_sl4x(system.file("extdata/in", "EXP1.sl4", 
                                package = "GTAPViz"))
# Get Data by Variable Name
sl4_data1 <- HARplus::get_data_by_var(c("qgdp", "EV"),sl4_data1)

# Rename variables in a dataset
rename_map <- data.frame(OldName = c("qgdp", "EV"), 
                        NewName = c("Real GDP", "Welfare"))
gtap_data <- rename_value(sl4_data1, column_name = "Variable", 
                          mapping.file = rename_map)

}
\seealso{
\code{\link{add_mapping_info}}, \code{\link{convert_units}}, \code{\link{sort_plot_data}}
}
\author{
Pattawee Puangchit
}
