% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{rename_value}
\alias{rename_value}
\title{Rename Values in Data Structures}
\usage{
rename_value(data, column_name = NULL, mapping.file)
}
\arguments{
\item{data}{Data structure containing data to rename}

\item{column_name}{Column name to modify. If NULL, extracted from mapping file}

\item{mapping.file}{Data frame with OldName and NewName columns for renaming}
}
\value{
A modified data frame or list of data frames with specified values replaced.
}
\description{
Replaces specified values in a column across nested data structures
}
\examples{
\dontrun{
# Example mapping file
wefare.decomp.rename <- data.frame(
  OldName = c("alloc_A1", "ENDWB1", "tech_C1"),
  NewName = c("Allocation", "Endowment", "Technology"),
  ColumnName = "Variable",
  stringsAsFactors = FALSE
)

# Load example data
har_data <- HARplus::load_harx(system.file("extdata/in", "EXP1.sl4",
                                          package = "GTAPViz"))

# Apply renaming
modified_data <- rename_value(har_data, mapping.file = wefare.decomp.rename)
}

}
\author{
Pattawee Puangchit
}
