% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{rename_value}
\alias{rename_value}
\title{Change Specific Values in a Dataframe Column}
\usage{
rename_value(data, column_name, original_values, new_values)
}
\arguments{
\item{data}{A data frame containing the column to modify.}

\item{column_name}{Character. The name of the column where values will be changed.}

\item{original_values}{Character vector. The values to be replaced.}

\item{new_values}{Character vector. The new values corresponding to original_values.}
}
\value{
A modified data frame with specified values replaced.
}
\description{
Replaces specified values in a given column with new values.
Only values listed in the original_values parameter will be changed.
}
\examples{
har_data <- HARplus::load_harx(c("globalcgds", "vgdpwld", "WEV"))
har_data <- HARplus::get_data_by_var(c("globalcgds", "vgdpwld", "WEV"), har_data)
modified_data <- rename_value(har_data[["A"]], "Variable",
                               c("alloc_A1", "ENDWB1", "tech_C1"),
                               c("Allocation", "Endowment", "Technology"))

}
\author{
Pattawee Puangchit
}
