% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{rename_value}
\alias{rename_value}
\title{Rename Values in a Data Frame or List of Data Frames}
\usage{
rename_value(data, column_name = NULL, mapping.file)
}
\arguments{
\item{data}{A data frame or a list of data frames containing the column to modify.}

\item{column_name}{Optional. Character. The name of the column where values will be changed.
If NULL, the function will attempt to extract \code{ColumnName} from \code{mapping.file}.}

\item{mapping.file}{A data frame containing \code{OldName} and \code{NewName} columns that specify
the mapping of values to be replaced. Optionally, a \code{ColumnName} column may define which
column should be modified.}
}
\value{
A modified data frame or list of data frames with specified values replaced.
}
\description{
Replaces specified values in a given column with new values based on a mapping file.
If \code{column_name} is not provided, it will be automatically extracted from \code{mapping.file} if available.
}
\examples{
\dontrun{
# Example mapping file
wefare.decomp.rename <- data.frame(
  OldName = c("alloc_A1", "ENDWB1", "tech_C1"),
  NewName = c("Allocation", "Endowment", "Technology"),
  ColumnName = "Variable",
  stringsAsFactors = FALSE
)

# Load example data
har_path <- system.file("extdata", "TAR10-WEL.har", package = "HARplus")
har_data <- HARplus::load_harx(har_path)

# Apply renaming
modified_data <- rename_value(har_data, mapping.file = wefare.decomp.rename)
}
}
\author{
Pattawee Puangchit
}
